fixed groupByResourceAndLibrary return value builder
In SpanAdapter.swift in the OpenTelemetryProtocol Exporter.

The objects being inserted into the result were not actually correctly populating, resulting in no instrumentationLibrarySpans being created.
This issue can be observed by stepping through the existing groupByResourceAndLibrary method. vars such as 'libraryInfo' will be updated, but the values in the underlying 'result' structure will not be.

added test to exercise the method.