added auto-delegate inject
This reconfigures the DDURLSessionDelegate object to be injected as an interception delegate, while holding the original delegate passed to the URLSession init method.
Externally, it will appear as wrapped object through the standard "isKindOfClass" checks, and will forward unimplemented methods to the wrapped delegate.
It seems like this will only capture data for dataTasks, so the swizzling needs to be extended to the other session task types, but this change wont impede their usage.
Additionally, if we decide to bring this into the instrumentation we should make it optional through configuration.