Ensure SemanticConvention script doesn't break tests
At the moment, if you run the `generate` script, it will make changes to the project which will cause compiler or test failures.

The three failures which this protects against:
* The value "int" being used as an enum value. This needs to be "Int" instead.
* The keyword "internal" being used as a case value. This should be wrapped with backticks like `internal`.
* The Swift telemetry language which we have chosen to add, outside of the specification.

While not technically a failure, this change will also revert the documentation changes made in a previous PR which changed all the `Int` values to "int" incorrectly. This fixes that too.